backend/requirements.txt
django
djangorestframework
python-dotenv==0.21.1